+++ src/util.cpp
@@ -93,7 +93,7 @@ void Util::ParseArgs(int argc,
   app.add_flag("--no-audio-device", args.no_audio_device,
                "Do not use audio device");
   app.add_flag("--list-devices", args.list_devices,
-               "List available video devices and exit");
+               "List available audio and video devices and exit");
 #if defined(USE_FAKE_CAPTURE_DEVICE)
   app.add_flag("--fake-capture-device", args.fake_capture_device,
                "Use fake video capture device instead of real camera");
@@ -128,6 +128,14 @@ void Util::ParseArgs(int argc,
   app.add_option("--video-input-device", args.video_device,
                  "Use the video input device specified by a name "
                  "(some device will be used if not specified)");
+#endif
+#if defined(__APPLE__)
+  app.add_option("--audio-input-device", args.audio_input_device,
+                 "Use the audio input device specified by an index or a name "
+                 "(use the system default if not specified)");
+  app.add_option("--audio-output-device", args.audio_output_device,
+                 "Use the audio output device specified by an index or a name "
+                 "(use the system default if not specified)");
 #endif
   app.add_option("--resolution", args.resolution,
                  "Video resolution (one of QVGA, VGA, HD, FHD, 4K, or "